% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghourglass-package.R
\docType{package}
\name{gghourglass-package}
\alias{gghourglass}
\alias{gghourglass-package}
\title{gghourglass: Plot Records per Time of Day}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Splits date and time of day components from continuous 'datetime' objects, then plots them using grammar of graphics ('ggplot2'). Plots can also be decorated with solar cycle information (e.g., sunset, sunrise, etc.). This is useful for visualising data that are associated with the solar cycle.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/gghourglass/}
  \item \url{https://www.github.com/pepijn-devries/gghourglass}
  \item Report bugs at \url{https://github.com/pepijn-devries/gghourglass/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Other contributors:
\itemize{
  \item Sander Lagerveld (\href{https://orcid.org/0000-0003-1291-4021}{ORCID}) [data contributor]
}

}
\keyword{internal}
